% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cert2idrssd.R
\name{cert2idrssd}
\alias{cert2idrssd}
\title{Convert bank identifier from CERT to IDRSSD}
\usage{
cert2idrssd(CERT)
}
\arguments{
\item{CERT}{An integer specifying the CERT number of the bank.}
}
\value{
An integer specifying the IDRSSD number of the bank. Returns
  \code{NULL} if there is an error or the FDIC API is unreachable.
}
\description{
This function takes a bank's CERT number as input and returns the corresponding IDRSSD number.
}
\examples{
\donttest{
cert2idrssd(3850)
}
}
