% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInstitutionsAll.R
\name{getInstitutionsAll}
\alias{getInstitutionsAll}
\title{Read FDIC Institution data set}
\usage{
getInstitutionsAll()
}
\value{
A data frame containing the FDIC Institution data set, or
  \code{NULL} if the resource is unreachable or returns an error.
}
\description{
This function reads the FDIC Institution data set from a URL (FDIC listing
of all institutions) and returns it as a data frame.
}
\examples{
\donttest{
dataInstitutions <- getInstitutionsAll()
}
}
