% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states2URL.R
\name{states2URL}
\alias{states2URL}
\title{Converts a vector of state names to a URL compatible format}
\usage{
states2URL(vec)
}
\arguments{
\item{vec}{A vector of state names to be converted to URL-compatible format}
}
\value{
A string containing the state names in URL-compatible format
}
\description{
This function takes a vector of state names and converts it to a format that is compatible with URLs. The resulting string can be used as a filter for APIs or other web requests.
}
