% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTaxonomy.R
\name{dataTaxonomy}
\alias{dataTaxonomy}
\title{Taxonomy Data}
\usage{
dataTaxonomy(name)
}
\arguments{
\item{name}{the name of the taxonomy file to extract. Available taxonomy names: "institution","location","history","summary","failure","financial".}
}
\value{
a data frame containing the extracted taxonomy information
}
\description{
Extracts the taxonomy information for a given name
}
\references{
For more information on the FDIC API, visit https://banks.data.fdic.gov/.
}
