% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTaxonomy.R
\name{getTaxonomy}
\alias{getTaxonomy}
\title{Get taxonomy file from FDIC website}
\usage{
getTaxonomy(taxonomy)
}
\arguments{
\item{taxonomy}{The name of the taxonomy file to download (one of "institution_properties.yaml", "location_properties.yaml", "history_properties.yaml", "summary_properties.yaml", "failure_properties.yaml", or "risview_properties.yaml")}
}
\description{
This function takes the name of the YAML file containing data taxonomy as an input
and downloads it from the FDIC website, saving it to the local directory for later use.
}
