% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cert2idrssd.R
\name{idrssd2cert}
\alias{idrssd2cert}
\title{Convert bank identifier from IDRSSD to CERT}
\usage{
idrssd2cert(IDRSSD)
}
\arguments{
\item{IDRSSD}{An integer specifying the IDRSSD number of the bank.}
}
\value{
An integer specifying the CERT number of the bank. Returns NULL if there is an error.
}
\description{
This function takes a bank's IDRSSD number as input and returns the corresponding CERT number.
}
\examples{
idrssd2cert(37)
}
